+++ src/Form/types/FilterSelect/index.js
@@ -1,0 +1,3 @@
+import FilterSelect from './FilterSelect';
+
+export default FilterSelect;
